Fixed bug in GLM menu of mrVista.
Edited eventMenu.m line 231 (new copy in mrvista_development/) to give the correct order of arguments for multiVoxelUI. Specifically, inserted a [] as the second (ROI) argument, so that getCurScan is in the third (scan) argument position. Fixes error when calling menu item GLM > MultiVoxelUI > Current Scan, which will fatally error if the scan number is greater than the number of loaded ROIs. At least fixes it in Inplane, didnt check for other views.